feat: Created getId for user controller
Created the getId function for the user controller so the user can do a GET request of a specific user sending the id trough the params of the request. The function will return an object with the user info trough the response, or will send and error if it doesn't find the user.